Move body of Client._post to State
